% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkATAC.R
\name{bulkATAC}
\alias{bulkATAC}
\title{Filtering BAM files and call peaks; Generate metrics plot and statistics used to predict sample quality}
\usage{
bulkATAC(
  original_bamfile,
  outdir,
  path_to_samtools = NULL,
  path_to_sambamba = NULL,
  path_to_macs2 = NULL,
  path_to_bedtools = NULL
)
}
\arguments{
\item{original_bamfile}{The BAM file of sample to be predicted}

\item{outdir}{output directory of all results}

\item{path_to_samtools}{the intalled path of samtools e.g. "/tools/bin/samtools"}

\item{path_to_sambamba}{the intalled path of sambamba e.g. "/tools/bin/sambamba"}

\item{path_to_macs2}{the intalled path of macs2 e.g. "/tools/bin/macs2"}

\item{path_to_bedtools}{the intalled path of bedtools e.g. "/tools/bin/bedtools"}
}
\value{
null
}
\description{
Filtering BAM files and call peaks; Generate metrics plot and statistics used to predict sample quality
}
\examples{
\dontrun{
library(bulkATACquality)
library(ATACseqQC)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)
library(phastCons100way.UCSC.hg19)
library(ChIPpeakAnno)
library(dplyr)
bamfile<-system.file("extdata","ex0001.bam",package = "bulkATACquality")
outdir<-substr(basename(bamfile),1,6)
dir.create(outdir)
bulkATAC(bamfile,outdir)
}
}
