% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPath.R
\name{.checkPath}
\alias{.checkPath}
\title{Check whether tools exist}
\usage{
.checkPath(u = NULL, path = NULL, throwError = TRUE)
}
\arguments{
\item{u}{name of tools}

\item{path}{path of tools}

\item{throwError}{whether throw errors when cannot find tools}
}
\value{
out True: find this tools in your machine; False: cannot find this tools in your machine;
}
\description{
Check whether tools exist
}
\examples{
.checkPath("samtools", throwError = FALSE)
}
