% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.R
\name{down}
\alias{down}
\title{downsampling bam files and generate metrics when downsampling 6000000, 9000000, 12000000,
15000000, 16000000, 17000000, 18000000, 19000000, 20000000 reads, see how metrics change}
\usage{
down(
  bam,
  outdir,
  path_to_samtools = NULL,
  path_to_sambamba = NULL,
  path_to_macs2 = NULL,
  path_to_bedtools = NULL
)
}
\arguments{
\item{bam}{bamfile to be downsampled}

\item{outdir}{directory of output files}

\item{path_to_samtools}{the intalled path of samtools e.g. "/tools/bin/samtools"}

\item{path_to_sambamba}{the intalled path of sambamba e.g. "/tools/bin/sambamba"}

\item{path_to_macs2}{the intalled path of macs2 e.g. "/tools/bin/macs2"}

\item{path_to_bedtools}{the intalled path of bedtools e.g. "/tools/bin/bedtools"}
}
\value{
null
}
\description{
downsampling bam files and generate metrics when downsampling 6000000, 9000000, 12000000,
15000000, 16000000, 17000000, 18000000, 19000000, 20000000 reads, see how metrics change
}
\examples{
\dontrun{
down(bam,"/project/exa_data/")
}
}
