% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAL.R
\name{GAl}
\alias{GAl}
\title{shifted all reads in input bamfile for downstream analysis, such as peak-calling and footprinting
The function shiftGAlignmentsList can be used to shift the reads. By default, all reads aligning to the positive strand are offset by +4bp,
and all reads aligning to the negative strand are offset by -5bp. The adjusted reads will be written into a new bamfile for peak calling
or footprinting.}
\usage{
GAl(bamfile)
}
\arguments{
\item{bamfile}{of sample to be shifted and predicted}
}
\value{
shifted alignment gal1
}
\description{
shifted all reads in input bamfile for downstream analysis, such as peak-calling and footprinting
The function shiftGAlignmentsList can be used to shift the reads. By default, all reads aligning to the positive strand are offset by +4bp,
and all reads aligning to the negative strand are offset by -5bp. The adjusted reads will be written into a new bamfile for peak calling
or footprinting.
}
\examples{
library(lattice)
library(ATACseqQC)
library(Rsamtools)
\dontrun{
bamfile<-system.file("extdata","exp.bam",package = "bulkATACquality")
gAl<-GAl(bamfile)
}
}
