% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHTML.R
\name{generateHTML}
\alias{generateHTML}
\title{Generate HTML of sample quality prediction}
\usage{
generateHTML(dir)
}
\arguments{
\item{dir}{folder name of all results (should be same as outdir of previous function)}
}
\value{
null
}
\description{
Generate HTML of sample quality prediction
}
\examples{
\dontrun{
generateHTML("/project/RL0001/")
}
}
