% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfr_result.R
\name{nfr_result}
\alias{nfr_result}
\title{Generate PT score plot and fit a line of all points
Generate density plot (hexbins) of points to visualize the distribution of points
Count the number of points with high coverage and high NFR score
Calculate variance and cov of hexbins}
\usage{
nfr_result(nfr, outliers, dir)
}
\arguments{
\item{nfr}{NFR object}

\item{outliers}{True —— including all points ; False —— excluding outliers}

\item{dir}{output directory}
}
\value{
p value of fitted line (whether the slope is significantly bigger than 0);
        counts of points with high coverage and high NFR score;
}
\description{
Generate PT score plot and fit a line of all points
Generate density plot (hexbins) of points to visualize the distribution of points
Count the number of points with high coverage and high NFR score
Calculate variance and cov of hexbins
}
\examples{
\dontrun{
NFR_T<-nfr_result(nfr,T)
p_NFR<-NFR_T$p_nfr
num_passingpoint<-NFR_T$passpoints_NFR
count_hexbins<- NFR_T$num_hexbins
var_countinhexbins<-NFR_T$var_hexbins
cov_countinhexbins<-NFR_T$cov_hexbins

p_NFR_excout<-nfr_result(nfr,F,dir0)
}
}
