% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_result.R
\name{pt_result}
\alias{pt_result}
\title{Generate PT score plot and fit a line of all points}
\usage{
pt_result(pt, outliers, dir)
}
\arguments{
\item{pt}{generated by ATACseqQC package and contains mean coverage and PT score of each point}

\item{outliers}{True —— including all points ; False —— excluding outliers}

\item{dir}{dir output directory}
}
\value{
p value of fitted line (whether the slope is significantly bigger than 0)
}
\description{
Generate PT score plot and fit a line of all points
}
\examples{
\dontrun{
p_PT<-pt_result(pt, outliers=TRUE, dir0)
p_PT_excout<-pt_result(pt, outliers=FALSE, dir0)
}
}
