% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS_signal.R
\name{TSS_signal}
\alias{TSS_signal}
\title{Generate TSS signal plot of nucleosome-bound region and nucleosome-free region
Calculate signal difference between the highest peak and the second highest peak
Calculate cov of signal}
\usage{
TSS_signal(bamfile, gal1, dir)
}
\arguments{
\item{bamfile}{path of filtered bamfile}

\item{gal1}{GAl object}

\item{dir}{output directory}
}
\value{
signal difference between the highest peak and the second highest peak and cov of signal
}
\description{
Generate TSS signal plot of nucleosome-bound region and nucleosome-free region
Calculate signal difference between the highest peak and the second highest peak
Calculate cov of signal
}
\examples{
\dontrun{
TSSsignal_result<-TSS_signal(bamfile,gAl,dir0)
TSSsignal_cov<-TSSsignal_result$cov_tsssignal
TSSsignal_difference<-TSSsignal_result$TSS_diff
}
}
