% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSE.R
\name{TSSE}
\alias{TSSE}
\title{Generate TSSE score plot and calculate width of curve and cov of score of all points}
\usage{
TSSE(dir, tsse)
}
\arguments{
\item{dir}{output directory}

\item{tsse}{TSSE object}
}
\value{
width of curve at mid-height and cov of score of all points
}
\description{
Generate TSSE score plot and calculate width of curve and cov of score of all points
}
\examples{
\dontrun{
TSSE_result<-TSSE(dir0,tsse)
TSSE_cov<-TSSE_result$cov_TSSscore
TSSE_wid<-TSSE_result$wid
}
}
